* Parallel/CommLayer.cpp (~CommLayer): Fix a printf format.
git-svn-id: https://svn.bcgsc.ca/svn/personal/jsimpson/partitionAssembly/PathWalker/src@39326 ad02020f-2a26-4133-947e-1829317c1416

+++ Parallel/CommLayer.cpp
@@ -17,7 +17,8 @@ CommLayer::CommLayer(int id, int kmerSize) : m_id(id), m_kmerSize(kmerSize), m_m
 //
 CommLayer::~CommLayer()
 {
-	printf("%d: Sent %llu messages\n", m_id, m_msgID);
+	printf("%d: Sent %llu messages\n", m_id,
+			(unsigned long long)m_msgID);
 	delete [] m_buffer;
 }
